backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

anthropic==0.40.0

sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

redis==5.2.1
hiredis==3.1.0

qdrant-client==1.12.2
voyageai==0.3.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

lxml==5.3.0
defusedxml==0.7.1
pymupdf==1.25.3
openpyxl==3.1.5

httpx==0.28.1
beautifulsoup4==4.12.3
pypdf2==3.0.1
pdfplumber==0.11.4
feedparser==6.0.11

celery[redis]==5.4.0
celery[gevent]==5.4.0

weasyprint==62.3
jinja2==3.1.4
stripe==11.4.1
psycopg2-binary==2.9.9

slowapi==0.1.9
sentry-sdk[fastapi]==2.19.2

pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
